apps/adk-agent/pyproject.toml
[project]
name = "mirothinker-adk"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "miroflow-tools>=0.1.0",
    "python-dotenv",
    "litellm",
]

[tool.uv.sources]
miroflow-tools = { path = "../../libs/miroflow-tools", editable = true }
